% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{clean_corpus}
\alias{clean_corpus}
\title{clessnverse::clean_corpus}
\usage{
clean_corpus(txt_bloc)
}
\arguments{
\item{txt_bloc}{blah}
}
\value{
return
}
\description{
clesns a bloc of text by removing double spaces, non-brekeable
spaces and L, S, D apostrophe (in french language).  This is particularly
}
\examples{
# To be documented

}
