% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\docType{data}
\name{clessn_palettes}
\alias{clessn_palettes}
\title{Colour palettes inspired by CLESSN branding}
\format{
An object of class \code{list} of length 1.
}
\usage{
clessn_palettes
}
\description{
Colour palettes inspired by CLESSN branding
}
\keyword{datasets}
