% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clessnverse-package.R
\docType{package}
\name{clessnverse-package}
\alias{clessnverse}
\alias{clessnverse-package}
\title{clessnverse: Package for Data Domestication, Analysis, and Visualization}
\description{
A repository of useful functions for the workflow of the CLESSN research group.
}
\author{
\strong{Maintainer}: William Poirier \email{william.poirier.1@ulaval.ca} (\href{https://orcid.org/0000-0002-3274-1351}{ORCID})

Authors:
\itemize{
  \item Alexandre Fortier-Chouinard \email{alexandre.fortier.chouinard@mail.utoronto.ca} (\href{https://orcid.org/0000-0002-6282-9709}{ORCID})
  \item Nadjim Fréchet \email{nadjim.frechet@umontreal.ca}
  \item Patrick Poncet \email{patrick@infoscope.ca}
  \item Judith Bourque \email{judith.bourque.2@ulaval.ca}
}

Other contributors:
\itemize{
  \item Yannick Dufresne \email{yannick.dufresne@pol.ulaval.ca} [copyright holder]
}

}
\keyword{internal}
