% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{commit_lake_item}
\alias{commit_lake_item}
\title{clessnverse::commit_lake_item}
\usage{
commit_lake_item(data, metadata, mode, credentials)
}
\description{
adds or replaces an object in the lake in a specific path with a specific key
}
\examples{
example
}
