% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{commit_mart_row}
\alias{commit_mart_row}
\title{Write a row in a data table of a CLESSN data mart}
\usage{
commit_mart_row(
  table_name,
  key,
  row = list(),
  refresh_data = FALSE,
  credentials
)
}
\arguments{
\item{table_name}{The name of the data mart table to write an observation to
without the 'chlub_tables_mart' prefix.}

\item{key}{A character string containing the unique primary key of this
observation in the table.  Data integrity of the CLESSN data
model is maintained by having a unique key per observation in
each table.}

\item{row}{A named list containing the observation to write to the datamart
table.  The names of the list \emph{are the columns} of the table.}

\item{refresh_data}{Logical.}

\item{credentials}{The hublot credentials obtained from the
hublot::get_credentials function}
}
\value{
returns a dataframe containing the data warehouse table with a JSON
attribute as well as a document.id and creation & update time stamps
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# connect to hublot
credentials <- hublot::get_credentials(
  Sys.getenv("HUB3_URL"),
  Sys.getenv("HUB3_USERNAME"),
  Sys.getenv("HUB3_PASSWORD")
  )


clessnverse::commit_mart_row(
  table_name = "political_parties_press_releases_freq",
  key = "QS212022",
  row = list(week_num=21, count=6, political_party="QS"),
  mode = "refresh",
  credentials = credentials)
}

}
