% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_l.R
\name{commit_warehouse_row}
\alias{commit_warehouse_row}
\title{clessnverse::commit_warehouse_row}
\usage{
commit_warehouse_row(table, key, row = list(), refresh_data, credentials)
}
\arguments{
\item{table}{be documented}

\item{key}{be documented}

\item{row}{be documented}

\item{refresh_data}{be documented}

\item{credentials}{be documented}
}
\value{
return
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Adds or replaces a rown in a warehouse with a specific key
}
\examples{
# To be documented
}
