% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_l.R
\name{commit_warehouse_table}
\alias{commit_warehouse_table}
\title{clessnverse::commit_warehouse_table}
\usage{
commit_warehouse_table(table_name, df, key_column, mode, credentials)
}
\arguments{
\item{to}{be documented}
}
\value{
to be documented
}
\description{
adds or replaces a rown in a warehouse with a specific key
}
\examples{
example
}
