% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{compute_category_sentiment_score}
\alias{compute_category_sentiment_score}
\title{clessnverse::compute_category_sentiment_score}
\usage{
compute_category_sentiment_score(
  txt_bloc,
  category_dictionary,
  sentiment_dictionary
)
}
\arguments{
\item{txt_bloc}{: the bloc of text to study}

\item{category_dictionary}{: a topic dictionary containing the categories to calculate the sentiment on}

\item{sentiment_dictionary}{: sentiment lexicoder dictionary}
}
\value{
returns a dataframe containing the sentiment score of each category in the category dictionary
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

calculates the sentiment score using a word stems dictionnary by category
}
\examples{
# To be documented

}
