% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{compute_catergory_sentiment_score}
\alias{compute_catergory_sentiment_score}
\title{clessnverse::commit_mart_table}
\usage{
compute_catergory_sentiment_score(
  txt_bloc,
  category_dictionary,
  sentiment_dictionary
)
}
\arguments{
\item{txt_bloc}{: the bloc of text to study}

\item{category_dictionary}{: a topic dictionary containing the categories to calculate the sentiment on}

\item{sentiment_dictionary}{: sentiment lexicoder dictionary}
}
\value{
returns a dataframe containing the sentiment score of each category in the category dictionary
}
\description{
adds or replaces a table in a datamart with a specific key
}
\examples{
# To be documented

}
