% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{compute_nb_sentences}
\alias{compute_nb_sentences}
\title{clessnverse::compute_nb_sentences}
\usage{
compute_nb_sentences(txt_bloc)
}
\arguments{
\item{txt_bloc}{be documented}
}
\value{
return
}
\description{
calculates the number of sentences in a bloc of text
}
\examples{
# To be documented
}
