% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{compute_relevance_score}
\alias{compute_relevance_score}
\title{clessnverse::compute_relevance_score}
\usage{
compute_relevance_score(txt_bloc, dictionary)
}
\value{
the function returns
}
\description{
calculates the relevance of a bloc of text against a topic dictionary
}
\examples{
example
}
