% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{count_na}
\alias{count_na}
\title{Count NA in a vector}
\usage{
count_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
number of NA in \code{x} (integer)
}
\description{
Count NA in a vector
}
\examples{
x <- c(4, 6, NA, 3, NA, 1)
count_na(x)

z <- c(NA, NA, "w", "a", "b", NA)
count_na(z)

}
