% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{countSentences}
\alias{countSentences}
\title{clessnverse::countSentences}
\usage{
countSentences(textblock)
}
\arguments{
\item{textblock}{: the text to analyse}
}
\value{
sentence.count : the number of sentences in the block
}
\description{
counts and returns the numbers of sentences in a text block
}
\examples{
example


}
