% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{countVecSentences}
\alias{countVecSentences}
\title{clessnverse}
\usage{
countVecSentences(vecCorpus)
}
\arguments{
\item{vecCorpus}{the vector of text strings}
}
\value{

}
\description{
counts and returns the umber of sentences in a vector of text blocks (strings)
}
\examples{
example
}
