% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dataviz}
\alias{theme_clean_dark}
\title{Data Visualization}
\usage{
theme_clean_light(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  primary_colour = "black",
  secondary_colour = "grey30",
  minor_colour = "#f7f7f7",
  bg_colour = "white"
)

theme_clean_dark(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  half_line = base_size/2,
  primary_colour = "white",
  secondary_colour = "#f2f2f2",
  minor_colour = "#525252",
  bg_colour = "#494949"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{primary_colour}{title and axis colour}

\item{secondary_colour}{text and axis ticks colour}

\item{minor_colour}{major gridlines colour}

\item{bg_colour}{plot background colour}
}
\description{
Functions for data visualization.
}
\details{
\describe{

\item{\code{theme_clean_light()}}{
A clean-looking ggplot2 theme, with x axis line and y major gridlines on a white background.}

\item{\code{theme_clean_dark()}}{
A clean-looking ggplot2 theme, with x axis line and y major gridlines or axis ticks on a dark grey background.}
}
}
\examples{
\dontrun{
p  <- ggplot2::ggplot(data = ggplot2::mpg) +
 ggplot2::geom_point(mapping = ggplot2::aes(x = displ, y = cty, colour = class)) +
 ggplot2::labs(
   title = "Un très très beau graphique informatif",
   subtitle = "Un sous-titre explicatif qui en dit long sur ces données importantes",
   caption = "Données: API Twitter \nCLESSN"
 ) +
 ggplot2::xlab("x axis label") +
ggplot2::ylab("y axis label")

p + theme_clean_light()
p + theme_clean_dark()
}
}
