% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_dropbox.R
\name{dbxDeleteFile}
\alias{dbxDeleteFile}
\title{clessnverse::dbxDeleteFile}
\usage{
dbxDeleteFile(filename, token)
}
\arguments{
\item{filename}{Filename to delete}

\item{token}{Token to pass on to 'httr'}
}
\description{
clessnverse::dbxDeleteFile
}
