% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_dropbox.R
\name{dbxListDir}
\alias{dbxListDir}
\title{clessnverse::dbxListDir}
\usage{
dbxListDir(dir, token)
}
\arguments{
\item{dir}{Directory}

\item{token}{Token to pass on to 'httr'}
}
\description{
clessnverse::dbxListDir
}
