% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_dropbox.R
\name{dbxMoveFile}
\alias{dbxMoveFile}
\title{clessnverse::dbxMoveFile}
\usage{
dbxMoveFile(source, destination, token, overwrite = FALSE)
}
\arguments{
\item{source}{Source path}

\item{destination}{Destination path}

\item{token}{Token to pass on to 'httr'}

\item{overwrite}{Logical. If overwrite == TRUE, then overwrites file in the destination.}
}
\description{
clessnverse::dbxMoveFile
}
