% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{detect_language}
\alias{detect_language}
\title{clessnverse::detect_language}
\usage{
detect_language(engine, text)
}
\arguments{
\item{engine}{: "azure" | "deeptranslate" | "fastText"}

\item{text}{: the text to translate}
}
\value{
2 chars language detected
}
\description{
detects the language of the text provided as a parameter
}
\examples{
example

}
