% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clessnverse.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Attach Hook Function}
\usage{
.onAttach(libname, pkgname)
}
\description{
This function is automatically called when the \code{clessnverse} package is attached using \code{library(clessnverse)}.
It ensures that the \code{sondr} package is also loaded, making all its functions available in the user's session.
Additionally, it displays a startup message with a disclaimer about the maintenance status of \code{clessnverse}.
}
\keyword{internal}
