% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Retrieves a dictionary from hublot.}
\usage{
get_dictionary(topic, lang = c("en", "fr"), credentials)
}
\arguments{
\item{topic}{The name or topic of the dictionary to retrieve from hublot.}

\item{lang}{The language of the dictionary. "en" is for English,
"fr" is for French. Both are included by default.}

\item{credentials}{The user's personal credentials from hublot.}
}
\value{
A quanteda type dictionary object.
}
\description{
Creates a dictionary object from a dictionary located in
the CLESSN data lake (hublot).
}
\examples{

\dontrun{
 # get credentials from hublot
 credentials <- hublot::get_credentials(
   Sys.getenv("HUB3_URL"),
   Sys.getenv("HUB3_USERNAME"),
   Sys.getenv("HUB3_PASSWORD")
   )
 # retrieve the COVID dictionary in both EN and FR
 clessnverse::get_dictionary("covid", c("en", "fr"), credentials)
}
}
\author{
CLESSN
}
