% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_e.R
\name{get_europe_mep_data}
\alias{get_europe_mep_data}
\title{clessnverse::getEuropeMepData}
\usage{
get_europe_mep_data(mep_full_name)
}
\arguments{
\item{mep_full_name}{: the full name of the MEP to lookup}
}
\value{
a dataframe
}
\description{
retrieves attributes of a MEP in the european parliament
}
\examples{
example
}
