% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{get_mart_table}
\alias{get_mart_table}
\title{clessnverse::get_mart_table}
\usage{
get_mart_table(table, credentials, nbrows = 0)
}
\arguments{
\item{table}{: the table name to fetch data from in the hub}

\item{credentials}{: your hub credential token}

\item{nbrows}{: the number of rows to fetch from the table}
}
\value{
: a dataframe containing the columns and rows
}
\description{
This function allows to extract data from a datamart
}
\examples{
example


}
