% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{get_warehouse_table}
\alias{get_warehouse_table}
\title{clessnverse::get_warhouse_table}
\usage{
get_warehouse_table(table, credentials, nbrows = 0)
}
\arguments{
\item{table}{: The name of the table to retrieve from the warhouse without
the}

\item{credentials}{: The hublot credentials obtained from the hublot::}

\item{nbrows}{: If nbrows is greater than 0, the dataframe returned will be
limited to nbrows observations.  This is particularly useful
when trying to see if there are records in a table and what
how structured they are.}
}
\value{
returns a dataframe containing the data warhouse table with a JSON
attribute as well as a document.id and creation & update time stamps
}
\description{
get_warehouse_table allows the programmer to retrieve a data
table from the CLESSN data warehouse named hublot.
}
\examples{
example


}
