% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getDictionary}
\alias{getDictionary}
\title{clessnverse::getDictionary}
\usage{
getDictionary(topic = "covid", method = "wordmatch", language = "")
}
\value{
dataframe
}
\description{
clessnverse::getDictionary
}
\examples{



}
