% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{getMepData}
\alias{getMepData}
\title{clessnverse}
\usage{
getMepData(mep_full_name)
}
\arguments{
\item{mep_full_name}{: the full name of the MEP to lookup}
}
\value{
a dataframe
}
\description{
retrieves attributes of a MEP in the european parliament
}
\examples{
example
}
