% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{\%vcontains\%}
\alias{\%vcontains\%}
\title{\%vcontains\%}
\usage{
vector \%vcontains\% values
}
\arguments{
\item{vector}{: A vector containing all possible values}

\item{values}{: each individual value, in the form of a vector, to check the
presence of in 'vector'}
}
\value{
\itemize{
\item TRUE if all the values in the vector 'values' are contained in the
vector 'vector'
= FALSE if all the values in 'values' are not contained in 'vector'
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Check if a vector 'vector' contains all values specified in the
vector 'values'
}
\examples{
# To be documented

}
