% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{log_close}
\alias{log_close}
\title{clessnverse::log_close}
\usage{
log_close(logger)
}
\arguments{
\item{logger}{blah}
}
\value{
blah
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

blah
}
\examples{
# To be documented



}
