% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{logit}
\alias{logit}
\title{clessnverse::logit}
\usage{
logit(scriptname = "clessnverse", message = "", logger = NULL)
}
\arguments{
\item{scriptname}{blah}

\item{message}{blah}

\item{logger}{blah}
}
\value{
blah
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

desc
}
\examples{
# To be documented



}
