% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{process_command_line_options}
\alias{process_command_line_options}
\title{clessnverse::processCommandLineOptions}
\usage{
process_command_line_options()
}
\arguments{
\item{to}{be documented}
}
\value{
to be documented
}
\description{
Parse the command line options of the agora+ scrapers
}
\examples{
to be documented



}
