% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{reduce_outliers}
\alias{reduce_outliers}
\title{Reduce outliers with the interquartile range method}
\usage{
reduce_outliers(vector)
}
\arguments{
\item{vector}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Interquartile range method based off https://www.scribbr.com/statistics/outliers/
}
\examples{

# Create a random vector following a normal distribution
vector <- round(rnorm(20, mean = 5, sd = 3))
vector
hist(vector)

# Add outliers
vector[c(4,11)] <- c(-24, 32)
vector
hist(vector)

# Apply function
new_vector <- reduce_outliers(vector)
new_vector
hist(new_vector)
}
