% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{replace_null}
\alias{replace_null}
\title{replace_null}
\usage{
replace_null(x)
}
\arguments{
\item{x}{: A list pbject.  Can be a list of lists}
}
\value{
Returns the same list with NULL values replaced by NA
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Replaces NULL values by NA in a list of list.  This function
supporte uneven / unbalanced list of lists
}
\examples{
# To be documented

}
