% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run_dictionary}
\alias{run_dictionary}
\title{Calculate dictionary expression mentions in a text.}
\usage{
run_dictionary(data, text, dictionary, verbose = TRUE)
}
\arguments{
\item{data}{An object of type data.frame.}

\item{text}{The name of the character variable for which
dictionary expressions are to be matched against.}

\item{dictionary}{An object of type dictionary.}

\item{verbose}{TRUE or FALSE Whether activity messages are displayed on
the screen.}
}
\value{
A data.frame which includes one more column than
the number of dictionary categories. The first column is named
doc_id. Each other column is named after a dictionary category.
}
\description{
This function creates a data.frame which includes one column
with each observation's ID as well as one column for each
category of the inputted dictionary.
}
\examples{

\dontrun{

# Calculate the number of dictionary expression mentions in
a list of attitudes.

run_dictionary(data.frame(colnames(attitude)),
text = colnames(attitude),
dictionary = quanteda::data_dictionary_LSD2015)
}
}
\author{
CLESSN
}
