% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{runDictionary}
\alias{runDictionary}
\title{clessnverse::runDictionary}
\usage{
runDictionary(corpusA, dataA, word, dfmA, dataB, dictionaryA)
}
\value{
dataframe
}
\description{
clessnverse::runDictionary
}
\examples{



}
