% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{sample_biased}
\alias{sample_biased}
\title{Create samples biased on the categories of one variable.}
\usage{
sample_biased(data, x, probs, size, iterations = 1, replace = FALSE)
}
\arguments{
\item{data}{An object of type data.frame.}

\item{x}{A vector included in the data, which will be
used to bias the sample.}

\item{probs}{A vector of probabilities which has the
same length as the number of values x can take.}

\item{size}{The number of observations to include
in each sample.}

\item{iterations}{The number of samples to take from
the data.frame.}

\item{replace}{Should sampling be with replacement?}
}
\value{
The sampled data, with all columns from the
original data.frame and a \code{prob} column indicating
the probability of inclusion of each observation. If
\code{iterations > 1}, an \code{iterations} column is included
to identify which iteration the observation belongs to.
}
\description{
This function creates a data.frame which includes one
or more samples taken from the original data. These samples
are biased on the categories of one variable: some of the
categories can be given a lower probability of inclusion
than others.
}
\examples{

\dontrun{

# Create a 100-respondent sample where men are three
# times more likely to be included than women.

sample_biased(Data, genderWoman,
probs = c(3, 1), size = 100)
}

}
\author{
CLESSN
}
