% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_discrete_quorum}
\alias{scale_discrete_quorum}
\title{Apply a discrete scale}
\usage{
scale_discrete_quorum()
}
\description{
These functions allow you to apply the CLESSN's discrete scales.
}
\examples{
\dontrun{
p  <- ggplot2::ggplot(data = ggplot2::mpg) +
 ggplot2::geom_point(mapping = ggplot2::aes(x = displ, y = cty, colour = class)) +
 ggplot2::labs(
   title = "Look at this graph!",
   subtitle = "What a great look, eh?",
   caption = "Data: Twitter API \nCLESSN"
   )

 p + scale_discrete_quorum()
}

}
\author{
Judith Bourque
}
