% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{split_words}
\alias{split_words}
\title{clessnverse::splitWords}
\usage{
split_words(txt)
}
\arguments{
\item{txt}{the string}
}
\value{
a list of words
}
\description{
returns a list of words of a sentence
}
\examples{
example



}
