% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{spread_list_to_df}
\alias{spread_list_to_df}
\title{spread_list_to_df}
\usage{
spread_list_to_df(l)
}
\arguments{
\item{l}{: A list object can be a list of lists (nested lists) that can have
uneven geometries}
}
\value{
\itemize{
\item TRUE if all the values in the vector 'values' are contained in the
vector 'vector'
= FALSE if all the values in 'values' are not contained in 'vector'
}
}
\description{
converts uneven nested lists to a dataframe
}
\examples{
# To be documented

}
