% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{theme_clean_light}
\alias{theme_clean_light}
\title{CLESSN themes}
\usage{
theme_clean_light(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
These are themes that correspond to the CLESSN's branding and related
projects.
}
\details{
\describe{

\item{\code{theme_clean_light()}}{
A clean-looking ggplot2 theme, with x axis line and y major gridlines on a white background.}

\item{\code{theme_clean_dark()}}{
A clean-looking ggplot2 theme, with x axis line and y major gridlines or axis ticks on a dark grey background.}
}
}
