% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{translate_text}
\alias{translate_text}
\title{clessnverse::translate_text}
\usage{
translate_text(
  text,
  engine = "azure",
  source_lang = NA,
  target_lang = "en",
  translate = FALSE
)
}
\arguments{
\item{text}{: the text to translate}

\item{engine}{: "azure" | "deeptranslate"}

\item{source_lang}{: the language to translate from - if NA, then automatic detection will be performed}

\item{target_lang}{: which language to translate to}

\item{translate}{: translates for real.  this is helpful to set to FALSE when debugging to avoid translation API fees}
}
\value{
a vector containing 1. a string = language detected and 2. a string =  the translated text
}
\description{
translates the text provided as a parameter using language autodetection in the language
}
\examples{
example


}
