% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twittosphere.R
\name{twittosphere}
\alias{twittosphere}
\alias{get_twitter_accounts}
\title{Make a twittosphere}
\usage{
get_twitter_accounts(type = c("journalist", "media"))
}
\arguments{
\item{type}{'journalist' or 'media'.}
}
\value{
A dataframe containing twitter accounts infos.
}
\description{
Functions used to generate data to make a twittosphere
}
\examples{
get_twitter_accounts(type = "journalist")
get_twitter_accounts(type = "media")
}
