% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_proportions}
\alias{calculate_proportions}
\title{Calculate the proportion of each category from one variable.}
\usage{
calculate_proportions(data, variable)
}
\arguments{
\item{data}{An object of type data.frame.}

\item{variable}{The name of the variable from which to calculate
the proportions.}
}
\value{
A data.frame which includes 3 columns.
1) `variable`: a column containing the variable's categories;
2) `n`: a column containing each category's frequency;
3) `prop`: a column containing each category's proportion.
}
\description{
This function creates a data.frame which includes 3 columns.
1) a column containing the variable's categories;
2) a column containing each category's frequency;
3) a column containing each category's proportion.
}
\examples{

\dontrun{

# Calculate the proportions of each cylinder configuration
# from mtcars.

calculate_proportions(mtcars,cyl)
}
}
\author{
CLESSN
}
