% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_e.R
\name{commit_lake_item}
\alias{commit_lake_item}
\title{clessnverse::commit_lake_item}
\usage{
commit_lake_item(data, metadata, mode, credentials)
}
\arguments{
\item{data}{a named list object containing a key, a path and the item to be
added to the data lake}

\item{metadata}{a named list object containing the metadata to be applied to
the lake item}
}
\description{
adds or replaces an object in the lake in a specific path with
a specific key
}
\examples{
example
}
