% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{commit_mart_row}
\alias{commit_mart_row}
\title{clessnverse::commit_mart_row}
\usage{
commit_mart_row(table, key, row = list(), mode = "refresh", credentials)
}
\description{
adds or replaces a rown in a datamart with a specific key
}
\examples{
example
}
