% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{commit_mart_table}
\alias{commit_mart_table}
\title{clessnverse::commit_mart_table}
\usage{
commit_mart_table(table_name, df, key_column, mode, credentials)
}
\description{
adds or replaces a table in a datamart with a specific key
}
\examples{
example
}
