% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_l.R
\name{commit_warehouse_row}
\alias{commit_warehouse_row}
\title{clessnverse::commit_warehouse_row}
\usage{
commit_warehouse_row(table, key, row = list(), mode = "refresh", credentials)
}
\arguments{
\item{to}{be documented}
}
\value{
to be documented
}
\description{
adds or replaces a rown in a warehouse with a specific key
}
\examples{
example
}
