% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{compute_nb_words}
\alias{compute_nb_words}
\title{clessnverse::compute_nb_words}
\usage{
compute_nb_words(txt_bloc)
}
\arguments{
\item{txt_bloc}{be documented}
}
\value{
return
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the number of words in a bloc of text
}
\examples{
# To be documented
}
