% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{get_hub2_table}
\alias{get_hub2_table}
\title{clessnverse::get_hub2_table}
\usage{
get_hub2_table(table_name, filter = list(), max_pages = -1, hub_conf)
}
\arguments{
\item{table}{: The name of the table to retrieve from the warhouse without
the}

\item{credentials}{: The hublot credentials obtained from the hublot::}

\item{nbrows}{: If nbrows is greater than 0, the dataframe returned will be
limited to nbrows observations.  This is particularly useful
when trying to see if there are records in a table and what
how structured they are.}
}
\value{
returns a dataframe containing the data warhouse table with a JSON
attribute as well as a document.id and creation & update time stamps
}
\description{
get_hub2_table allows the programmer to retrieve a data
table from the CLESSN hub2 data warehouse.
}
\examples{
example


}
