% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{get_warehouse_table}
\alias{get_warehouse_table}
\title{clessnverse::get_warehouse_table}
\usage{
get_warehouse_table(table_name, credentials, data_filter = list(), nbrows = 0)
}
\arguments{
\item{table_name}{The name of the table to retrieve from the warehouse without
the 'chlub_tables_warehouse' prefix}

\item{credentials}{The hublot credentials obtained from the
hublot::get_credentials function}

\item{data_filter}{a filter on the data to be selected in the query}

\item{nbrows}{Optional argument
If nbrows is greater than 0, the dataframe returned will be
limited to nbrows observations.  This is particularly useful
when trying to see if there are records in a table and what
how structured they are.
If nbrows is omitted, then all rows of the table are returned}
}
\value{
returns a dataframe containing the data warehouse table content
with the document.id and creation & update time stamps
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

get_warehouse_table allows the programmer to retrieve a data
table from the CLESSN data warehouse named hublot.
}
\examples{
 \dontrun{
 # connect to hublot
 credentials <- hublot::get_credentials(
   Sys.getenv("HUB3_URL"),
   Sys.getenv("HUB3_USERNAME"),
   Sys.getenv("HUB3_PASSWORD")
   )

 # gets the entire warehouse table 'people'
 clessnverse::get_warehouse_table('people', credentials)

 # gets the first 10 rows of the 'political_parties_press_releases' table
 clessnverse::get_warehouse_table(
   table_name = 'political_parties_press_releases',
   data_filter = list(),
   credentials = credentials,
   nbrows=10
   )
 }

}
