% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{getDictionary}
\alias{getDictionary}
\title{clessnverse}
\usage{
getDictionary(topic = "covid", method = "wordmatch", language = "")
}
\arguments{
\item{topic}{: "covid" | "sentiment"}

\item{method}{: "wordmatch" | "regex" | "dfm"}

\item{language}{= "en" | "fr" | ""}
}
\value{
Formal class 'dictionary2' of package "quanteda" for method "dfm" or a vector of regex for method "regex"
}
\description{
retrives the clessn dictionnary related to a specific topic
}
\examples{
example



}
