% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{log_init}
\alias{log_init}
\title{clessnverse::log_init}
\usage{
log_init(script, backend, logpath = ".")
}
\arguments{
\item{script}{blah}

\item{backend}{blah}

\item{logpath}{blah}
}
\value{
return
}
\description{
This function initializes the log
}
\examples{
# To be documented



}
