% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{log_init}
\alias{log_init}
\title{clessnverse::log_init}
\usage{
log_init(script, backend, logpath = ".")
}
\arguments{
\item{to}{be documented}
}
\value{
to be documented
}
\description{
This function initializes the log
}
\examples{
example



}
