<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{logit}
\alias{logit}
\title{clessnverse::logit}
\usage{
logit(scriptname = "clessnverse", message = "", logger = NULL)
}
\description{
clessnverse::logit
}
\examples{
example



}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{logit}
\alias{logit}
\title{clessnverse::logit}
\usage{
logit(scriptname = "clessnverse", message = "", logger = NULL)
}
\arguments{
\item{scriptname}{blah}

\item{message}{blah}

\item{logger}{blah}
}
\value{
blah
}
\description{
desc
}
\examples{
# To be documented



}
>>>>>>> 268ebb0d38daa5a1a21e10c8fbcd6de7a736f96a
