% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{logit}
\alias{logit}
\title{clessnverse::logit}
\usage{
logit(scriptname = "clessnverse", message = "", logger = NULL)
}
\arguments{
\item{to}{be documented}
}
\value{
to be documented
}
\description{
to be documented
}
\examples{
example



}
