% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domestication.R
\name{normalize_min_max}
\alias{normalize_min_max}
\title{Normalize a continuous variable between 0 and 1}
\usage{
normalize_min_max(x, remove_na = T)
}
\arguments{
\item{vector}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Normalize a continuous variable between 0 and 1
}
\examples{
data <- tibble::tibble(a = c(2, 0, 0, 0), b = c(4, 0, 0, 0))

# Base R
data_output <- sapply(data, normalize_min_max)

# Dplyr
library("dplyr")

data_output <- data \%>\%
  mutate(across(c(a, b), normalize_min_max))
}
