% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{process_command_line_options}
\alias{process_command_line_options}
\title{clessnverse::process_command_line_options}
\usage{
process_command_line_options()
}
\value{
blah
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Parse the command line options of the agora+ scrapers
}
\examples{
# To be documented



}
