% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domestication.R
\name{reduce_outliers}
\alias{reduce_outliers}
\title{Reduce outliers with the interquartile range method}
\usage{
reduce_outliers(vector)
}
\arguments{
\item{vector}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Reduce outliers with the interquartile range method
}
