% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_t.R
\name{rm_accents}
\alias{rm_accents}
\title{Replace accented characters with their non-accented counterparts}
\usage{
rm_accents(str, pattern = "all")
}
\arguments{
\item{str}{: the text to remove accents from}

\item{pattern}{: patterns of the accents to remove}
}
\value{
str without accent
}
\description{
Replace accented characters with their non-accented counterparts
}
\examples{
example


}
