% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{runDictionary}
\alias{runDictionary}
\title{clessnverse::runDictionary}
\usage{
runDictionary(corpusA, dataA, word, dfmA, dataB, dictionaryA)
}
\arguments{
\item{corpusA}{the corpus}

\item{dataA}{the dataframe containing the words to check}

\item{word}{to be documented}

\item{dfmA}{to be documented}

\item{dataB}{to be documented}

\item{dictionaryA}{to be documented}
}
\value{
dataframe
}
\description{
Runs a dictionary against a text corpus and returns
}
\examples{
example



}
