% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{translateText}
\alias{translateText}
\title{clessnverse::translateText}
\usage{
translateText(text, engine = "azure", target_lang = "fr", fake = TRUE)
}
\arguments{
\item{text}{: the text to translate}

\item{engine}{: "azure" | "google"}

\item{target_lang}{: which language to translate to}
}
\value{
a string containing the translated text
}
\description{
translates the text provided as a parameter using language autodetection in the language
}
\examples{
example


}
